pyproject.toml
[project]
name = "learning-ai-nebius"
version = "0.1.0"
description = "Nebius AI Engineering course assignments"
requires-python = ">=3.11"
dependencies = [
    "litellm",
    "pydantic",
    "pandas",
    "openpyxl",
    "python-dotenv",
    "mlflow",
    "tqdm",
    "tenacity",
    "marimo>=0.22.4",
    "seaborn>=0.13.2",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
exclude = [".venv", "w2-ai-product/outputs"]

